+++ packages/iov-bcp-types/types/connection.d.ts
@@ -125,7 +125,7 @@ export interface BcpConnection {
     readonly disconnect: () => void;
     readonly chainId: () => ChainId;
     readonly height: () => Promise<number>;
-    readonly getTicker: (ticker: TokenTicker) => Promise<BcpQueryEnvelope<BcpTicker>>;
+    readonly getTicker: (ticker: TokenTicker) => Promise<BcpTicker | undefined>;
     readonly getAllTickers: () => Promise<ReadonlyArray<BcpTicker>>;
     /**
      * Get the current account information (e.g. balance)