+++ CHANGELOG.md
@@ -27,6 +27,8 @@ Breaking changes
   as we're migrating from wallet nicknames to username NFTs.
 * @iov/bcp-types: Convert `BcpQueryEnvelope` to `ReadonlyArray` in return
   type of `BcpConnection.getAllTickers`.
+* @iov/bcp-types: Convert `BcpQueryEnvelope` to `BcpTicker | undefined` in return
+  type of `BcpConnection.getTicker`.
 * @iov/core: Remove `MultiChainSigner.getNonce`. If you really need this, use
   `signer.connection(chainId).getNonce({ address: addr })` instead.
 * @iov/core: Removed BNS re-exports `bnsConnector`, `bnsFromOrToTag`,